LSTM cell implementation was transferred to common place. Unneccessary code was removed
