+++ python/tvm/relay/frontend/onnx.py
@@ -2142,351 +2142,6 @@ class RNN(OnnxOpConverter):
 class LSTM(RNN):
     """Operator converter for LSTM"""
 
-    @classmethod
-    def generate_lstm(
-        cls, X_steps, H_t, C_t, W, R, B, p_i, p_f, p_o, f_act, g_act, h_act, backwards=False
-    ):
-        """Create an unrolled lstm loop.
-
-        See https://github.com/onnx/onnx/blob/master/docs/Operators.md for math.
-        """
-        h_list = []
-        seq_length = len(X_steps)
-        for i in range(seq_length):
-            step = X_steps[i] if not backwards else X_steps[seq_length - (i + 1)]
-            step = _op.squeeze(step, axis=[0])
-            gates = _op.nn.dense(step, W) + _op.nn.dense(H_t, R)
-            if B is not None:
-                WB, RB = _op.split(B, 2)
-                gates += WB + RB
-            i, o, f, c = _op.split(gates, 4, axis=-1)
-
-            if p_i != 0:
-                i = f_act(i + p_i * C_t)
-            else:
-                i = f_act(i)
-
-            if p_f != 0:
-                f = f_act(f + p_f * C_t)
-            else:
-                f = f_act(f)
-
-            c = g_act(c)
-            C = f * C_t + i * c
-            if p_o != 0:
-                o = f_act(o + p_o * C)
-            else:
-                o = f_act(o)
-
-            H = o * h_act(C)
-
-            H_t = H
-            C_t = C
-            h_list.append(_op.expand_dims(H, axis=0))
-
-        if backwards:
-            # Canonical view is hidden states from the first token not last
-            h_list = h_list[::-1]
-
-        # Concatenate outputs and add back in direction axis.
-        concatenated = _op.concatenate(h_list, 0)
-        output = _op.expand_dims(concatenated, axis=1)
-        H_t = _op.expand_dims(H_t, axis=0)
-        C_t = _op.expand_dims(C_t, axis=0)
-
-        return output, H_t, C_t
-
-    @classmethod
-    def _impl_v7(cls, inputs, attr, params):
-        # Unpack inputs, note that if optional and not provided then value will be None.
-        X = inputs[0]
-        Wp = inputs[1]
-        Rp = inputs[2]
-        Bp = inputs[3]
-        # Sequence length currently unused as it can be inferred from shapes.
-        # sequence_lens = inputs['sequence_lens']
-        Hp_0 = inputs[5]
-        Cp_0 = inputs[6]
-        Pp = inputs[7]
-
-        num_directions = infer_shape(Wp)[0]
-        W_dtype = infer_type(Wp).checked_type.dtype
-
-        if num_directions not in [1, 2]:
-            raise ValueError("num_directions must be either 1 or 2!")
-
-        X_shape = infer_shape(X)
-        hidden_size = infer_shape(Rp)[-1]
-        batch_size = X_shape[1]
-
-        # Initialize state if not provided.
-        # Otherwise remove bidirectional axis.
-        if Hp_0 is None:
-            Hp_0 = _op.zeros((num_directions, batch_size, hidden_size), W_dtype)
-        if Cp_0 is None:
-            Cp_0 = _op.zeros((num_directions, batch_size, hidden_size), W_dtype)
-        if Bp is None:
-            Bp = _op.zeros((num_directions, hidden_size * 8), W_dtype)
-        if Pp is not None:
-            p_i, p_o, p_f = _op.split(Pp, 3, axis=1)
-        else:
-            p_i = p_o = p_f = _op.zeros((num_directions, hidden_size), W_dtype)
-
-        if "activations" in attr:
-            activations = attr["activations"]
-            if len(activations) != 3 * num_directions:
-                raise NotImplementedError(
-                    f"LSTM assumes 3 * num_directions activation functions are provided"
-                )
-            alpha_loc = 0
-            alphas = attr.get("activation_alpha", [])
-            if isinstance(alphas, float):
-                alphas = [alphas]
-            beta_loc = 0
-            betas = attr.get("activation_beta", [])
-            if isinstance(betas, float):
-                betas = [betas]
-            acts = []
-            for i in range(3 * num_directions):
-                alpha = None
-                beta = None
-                activation = activations[i]
-                if cls._activation_needs_alpha(activation) and len(alphas) > alpha_loc:
-                    alpha = alphas[alpha_loc]
-                    alpha_loc += 1
-                if cls._activation_needs_beta(activation) and len(betas) > beta_loc:
-                    beta = betas[beta_loc]
-                    beta_loc += 1
-                acts.append(cls._activation_helper(activation, alpha, beta))
-        else:
-            acts = [_op.sigmoid, _op.tanh, _op.tanh] * num_directions
-
-        X_steps = _op.split(X, indices_or_sections=X_shape[0], axis=0)
-        result_output = []
-        result_H = []
-        result_C = []
-
-        H_ts = _op.split(Hp_0, num_directions)
-        C_ts = _op.split(Cp_0, num_directions)
-        Ws = _op.split(Wp, num_directions)
-        Rs = _op.split(Rp, num_directions)
-        Bs = _op.split(Bp, num_directions)
-        p_is = _op.split(p_i, num_directions)
-        p_fs = _op.split(p_f, num_directions)
-        p_os = _op.split(p_o, num_directions)
-        for i in range(num_directions):
-            H_t = _op.squeeze(H_ts[i], axis=[0])
-            C_t = _op.squeeze(C_ts[i], axis=[0])
-            W = _op.squeeze(Ws[i], axis=[0])
-            R = _op.squeeze(Rs[i], axis=[0])
-            B = _op.squeeze(Bs[i], axis=[0])
-            p_i = _op.squeeze(p_is[i], axis=[0])
-            p_f = _op.squeeze(p_fs[i], axis=[0])
-            p_o = _op.squeeze(p_os[i], axis=[0])
-
-            f_act, g_act, h_act = acts[i * 3 : (i + 1) * 3]
-            output, H, C = LSTM.generate_lstm(
-                X_steps=X_steps,
-                H_t=H_t,
-                C_t=C_t,
-                W=W,
-                R=R,
-                B=B,
-                p_i=p_i,
-                p_f=p_f,
-                p_o=p_o,
-                f_act=f_act,
-                g_act=g_act,
-                h_act=h_act,
-                backwards=i == 1,
-            )
-
-            result_output.append(output)
-            result_H.append(H)
-            result_C.append(C)
-
-        output = _op.concatenate(result_output, axis=1)
-        H = _op.concatenate(result_H, axis=0)
-        C = _op.concatenate(result_C, axis=0)
-
-        return _expr.TupleWrapper(_expr.Tuple((output, H, C)), 3)
-
-
-class LSTM_dev(RNN):
-    """Operator converter for LSTM"""
-
-    @classmethod
-    def generate_lstm(
-        cls, X_steps, H_t, C_t, W, R, WB, RB, p_i, p_f, p_o, f_act, g_act, h_act, backwards=False
-    ):
-        """Create an unrolled lstm loop.
-
-        See https://github.com/onnx/onnx/blob/master/docs/Operators.md for math.
-        """
-        h_list = []
-        seq_length = len(X_steps)
-        for i in range(seq_length):
-            step = X_steps[i] if not backwards else X_steps[seq_length - (i + 1)]
-            step = _op.squeeze(step, axis=[0])
-            gates = _op.nn.dense(step, W) + _op.nn.dense(H_t, R)
-            if WB is not None and RB is not None:
-                gates += WB + RB
-            i, o, f, c = _op.split(gates, 4, axis=-1)
-            if p_i != 0:
-                i = f_act(i + p_i * C_t)
-            else:
-                i = f_act(i)
-
-            if p_f != 0:
-                f = f_act(f + p_f * C_t)
-            else:
-                f = f_act(f)
-
-            c = g_act(c)
-            C_t = f * C_t + i * c
-            if p_o != 0:
-                o = f_act(o + p_o * C_t)
-            else:
-                o = f_act(o)
-
-            H_t = o * h_act(C_t)
-
-            h_list.append(_op.expand_dims(H_t, axis=0))
-
-        if backwards:
-            # Canonical view is hidden states from the first token not last
-            h_list = h_list[::-1]
-
-        # Concatenate outputs and add back in direction axis.
-        concatenated = _op.concatenate(h_list, 0)
-        output = _op.expand_dims(concatenated, axis=1)
-        H_t = _op.expand_dims(H_t, axis=0)
-        C_t = _op.expand_dims(C_t, axis=0)
-
-        return output, H_t, C_t
-
-    @classmethod
-    def _impl_v7(cls, inputs, attr, params):
-        # Unpack inputs, note that if optional and not provided then value will be None.
-        X = inputs[0]
-        Wp = inputs[1]
-        Rp = inputs[2]
-        Bp = inputs[3]
-        # Sequence length currently unused as it can be inferred from shapes.
-        # sequence_lens = inputs['sequence_lens']
-        Hp_0 = inputs[5]
-        Cp_0 = inputs[6]
-        Pp = inputs[7]
-
-        num_directions = infer_shape(Wp)[0]
-        W_dtype = infer_type(Wp).checked_type.dtype
-
-        if num_directions not in [1, 2]:
-            raise ValueError("num_directions must be either 1 or 2!")
-
-        X_shape = infer_shape(X)
-        hidden_size = infer_shape(Rp)[-1]
-        batch_size = X_shape[1]
-
-        # Initialize state if not provided.
-        # Otherwise remove bidirectional axis.
-        if Hp_0 is None:
-            Hp_0 = _op.zeros((num_directions, batch_size, hidden_size), W_dtype)
-        if Cp_0 is None:
-            Cp_0 = _op.zeros((num_directions, batch_size, hidden_size), W_dtype)
-        if Bp is None:
-            Bp = _op.zeros((num_directions, hidden_size * 8), W_dtype)
-        if Pp is not None:
-            p_i, p_o, p_f = _op.split(Pp, 3, axis=1)
-        else:
-            p_i = p_o = p_f = _op.zeros((num_directions, hidden_size), W_dtype)
-
-        if "activations" in attr:
-            activations = attr["activations"]
-            if len(activations) != 3 * num_directions:
-                raise NotImplementedError(
-                    f"LSTM assumes 3 * num_directions activation functions are provided"
-                )
-            alpha_loc = 0
-            alphas = attr.get("activation_alpha", [])
-            if isinstance(alphas, float):
-                alphas = [alphas]
-            beta_loc = 0
-            betas = attr.get("activation_beta", [])
-            if isinstance(betas, float):
-                betas = [betas]
-            acts = []
-            for i in range(3 * num_directions):
-                alpha = None
-                beta = None
-                activation = activations[i]
-                if cls._activation_needs_alpha(activation) and len(alphas) > alpha_loc:
-                    alpha = alphas[alpha_loc]
-                    alpha_loc += 1
-                if cls._activation_needs_beta(activation) and len(betas) > beta_loc:
-                    beta = betas[beta_loc]
-                    beta_loc += 1
-                acts.append(cls._activation_helper(activation, alpha, beta))
-            f_act, g_act, h_act = acts
-        else:
-            f_act = _op.sigmoid
-            g_act = _op.tanh
-            h_act = _op.tanh
-
-        X_steps = _op.split(X, indices_or_sections=X_shape[0], axis=0)
-        result_output = []
-        result_H = []
-        result_C = []
-
-        H_ts = _op.split(Hp_0, num_directions)
-        C_ts = _op.split(Cp_0, num_directions)
-        Ws = _op.split(Wp, num_directions)
-        Rs = _op.split(Rp, num_directions)
-        Bs = _op.split(Bp, num_directions)
-        p_is = _op.split(p_i, num_directions)
-        p_fs = _op.split(p_f, num_directions)
-        p_os = _op.split(p_o, num_directions)
-        for i in range(num_directions):
-            H_t = _op.squeeze(H_ts[i], axis=[0])
-            C_t = _op.squeeze(C_ts[i], axis=[0])
-            W = _op.squeeze(Ws[i], axis=[0])
-            R = _op.squeeze(Rs[i], axis=[0])
-            WB, RB = _op.split(Bs[i], 2, -1)
-            p_i = _op.squeeze(p_is[i], axis=[0])
-            p_f = _op.squeeze(p_fs[i], axis=[0])
-            p_o = _op.squeeze(p_os[i], axis=[0])
-
-            output, H, C = LSTM_dev.generate_lstm(
-                X_steps=X_steps,
-                H_t=H_t,
-                C_t=C_t,
-                W=W,
-                R=R,
-                WB=WB,
-                RB=RB,
-                p_i=p_i,
-                p_f=p_f,
-                p_o=p_o,
-                f_act=f_act,
-                g_act=g_act,
-                h_act=h_act,
-                backwards=i == 1,
-            )
-
-            result_output.append(output)
-            result_H.append(H)
-            result_C.append(C)
-
-        output = _op.concatenate(result_output, axis=1)
-        H = _op.concatenate(result_H, axis=0)
-        C = _op.concatenate(result_C, axis=0)
-
-        return _expr.TupleWrapper(_expr.Tuple((output, H, C)), 3)
-
-
-class LSTM_dev2(RNN):
-    """Operator converter for LSTM"""
-
     # TODO (vvchernov): unbind was gotten from pytorch.py and modified.
     # It looks like torch.unbind
     # It needs such operation on relay side to avoid excess manipulation like squeeze
@@ -2502,67 +2157,6 @@ class LSTM_dev2(RNN):
         return ret
 
     @classmethod
-    def lstm_cell(
-        cls,
-        input_seqs,
-        H_t,
-        C_t,
-        Wi,
-        Wh,
-        Bi,
-        Bh,
-        P,
-        p_i,
-        p_f,
-        p_o,
-        f_act,
-        g_act,
-        h_act,
-        backwards=False,
-    ):
-        # Input hidden state shape = (batch, hidden_size)
-        # Wi, Wh, Bi, Bh, proj matrix P, peephole matrices: p_i, p_f, p_o are expected.
-        # Wi and Wh shoud exist the others can be None
-
-        outputs_list = []
-        for x_t in input_seqs if not backwards else reversed(input_seqs):
-            # x_t shape = (batch, feature size), step shape = (batch, feature size + hidden_size)
-            step = _op.concatenate([x_t, H_t], axis=1)
-            W = _op.concatenate([Wi, Wh], axis=1)
-            # Instead of _op.nn.dense(x_t, weights[0]) + _op.nn.dense(H_t, weights[1]) we have _op.nn.dense(step, W)
-            # gates shape = (batch, 4 * hidden_size)
-            gates = _op.nn.dense(step, W)
-            # Add biases
-            if Bi is not None:
-                gates += Bi
-            if Bh is not None:
-                gates += Bh
-            i, f, c, o = _op.split(gates, 4, axis=-1)  # (batch, hidden_size)
-
-            if p_i is not None and p_f is not None:
-                i = f_act(i + p_i * C_t)
-                f = f_act(f + p_f * C_t)
-            else:
-                i = f_act(i)
-                f = f_act(f)
-
-            c = g_act(c)
-            C_t = f * C_t + i * c
-            if p_o is not None:
-                o = f_act(o + p_o * C_t)
-            else:
-                o = f_act(o)
-
-            H_t = o * h_act(C_t)
-
-            if P is not None:
-                H_t = _op.nn.dense(H_t, P)
-
-            outputs_list.append(H_t)  # [seq_num, (batch, hidden_size)]
-
-        return outputs_list, H_t, C_t
-
-    @classmethod
     def bidir_lstm_cell(
         cls,
         input_seqs,
@@ -2572,7 +2166,7 @@ class LSTM_dev2(RNN):
         h_act,
     ):
         seq_len = len(input_seqs)
-        forward_outputs, fw_H_t, fw_C_t = LSTM_dev2.lstm_cell(
+        forward_outputs, fw_H_t, fw_C_t = _op.lstm_cell(
             input_seqs,
             **weight_dicts[0],
             f_act=f_act,
@@ -2580,7 +2174,7 @@ class LSTM_dev2(RNN):
             h_act=h_act,
         )
 
-        reverse_outputs, rev_H_t, rev_C_t = LSTM_dev2.lstm_cell(
+        reverse_outputs, rev_H_t, rev_C_t = _op.lstm_cell(
             input_seqs,
             **weight_dicts[1],
             f_act=f_act,
@@ -2663,8 +2257,8 @@ class LSTM_dev2(RNN):
             g_act = _op.tanh
             h_act = _op.tanh
 
-        # It can be replaced by _op.split if issue #8412 is resolved
-        X_steps = LSTM_dev2.unbind(X, axis=0)
+        # TODO (vvchernov): It can be replaced by _op.split if issue #8412 is resolved
+        X_steps = LSTM.unbind(X, axis=0)
 
         H_ts = _op.split(Hp_0, num_directions)
         C_ts = _op.split(Cp_0, num_directions)
@@ -2689,26 +2283,18 @@ class LSTM_dev2(RNN):
 
             weights_dict["Wi"] = _op.squeeze(Ws[i], axis=[0])
             weights_dict["Wh"] = _op.squeeze(Rs[i], axis=[0])
-            if Bp is None:
-                Bi = None
-                Bh = None
-            else:
+            if Bp is not None:
                 Bi, Bh = _op.split(Bs[i], 2, -1)
-            weights_dict["Bi"] = Bi
-            weights_dict["Bh"] = Bh
-            weights_dict["P"] = None
-            if Pp is None:
-                weights_dict["p_i"] = None
-                weights_dict["p_f"] = None
-                weights_dict["p_o"] = None
-            else:
+                weights_dict["Bi"] = Bi
+                weights_dict["Bh"] = Bh
+            if Pp is not None:
                 weights_dict["p_i"] = _op.squeeze(p_is[i], axis=[0])
                 weights_dict["p_f"] = _op.squeeze(p_fs[i], axis=[0])
                 weights_dict["p_o"] = _op.squeeze(p_os[i], axis=[0])
             weights_dicts.append(weights_dict)
 
         if num_directions == 2:
-            output, H, C = LSTM_dev2.bidir_lstm_cell(
+            output, H, C = LSTM.bidir_lstm_cell(
                 input_seqs=X_steps,
                 weight_dicts=weights_dicts,
                 f_act=f_act,
@@ -2717,7 +2303,7 @@ class LSTM_dev2(RNN):
             )
         else:
             # outputs shape = [seqs_num, (batch_size, hidden_size)]
-            outputs, H, C = LSTM_dev2.lstm_cell(
+            outputs, H, C = _op.lstm_cell(
                 input_seqs=X_steps,
                 **weights_dicts[0],
                 f_act=f_act,
@@ -3949,7 +3535,7 @@ def _get_convert_map(opset):
         "Flatten": Flatten.get_converter(opset),
         "LRN": LRN.get_converter(opset),
         # Recurrent Layers
-        "LSTM": LSTM_dev2.get_converter(opset),
+        "LSTM": LSTM.get_converter(opset),
         "GRU": GRU.get_converter(opset),
         # defs/vision
         "MaxRoiPool": MaxRoiPool.get_converter(opset),
