+++ python/tvm/relay/op/__init__.py
@@ -38,6 +38,7 @@ from .reduce import *
 from .tensor import *
 from .transform import *
 from .algorithm import *
+from .layers import *
 from . import vm
 from . import nn
 from . import annotation
